feat:上传图片成功后会保存记录到数据库
- 记录在 file 表

+++ domain/file/content.go
@@ -29,8 +29,8 @@ func calculateHash(data []byte) string {
 }
 
 // SizeInBytes 文件大小
-func (c *Content) SizeInBytes() uint64 {
-	return c.sizeInBytes
+func (c *Content) SizeInBytes() int64 {
+	return int64(c.sizeInBytes)
 }
 
 // Hash 文件哈希值
